+++ src/stores/index.ts
@@ -1,1 +1,1 @@
-export * from "./useUserStore";
+export * from "./useUserStore/useUserStore";
