+++ src/stores/useUserStore/types.ts
@@ -1,0 +1,11 @@
+import type { UserInfo } from "../../types";
+
+export interface UserStore {
+  isAuthenticated: boolean;
+  user: UserInfo | null;
+  token: string | null;
+  error: string | null;
+
+  setError: (error: string | null) => void;
+  clearError: () => void;
+}
